requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
cloudinary==1.29.0
dj-database-url==0.5.0
Django==3.1.3
django-cloudinary-storage==0.3.0
django-cors-headers==3.5.0
django-heroku==0.3.1
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
drf-yasg==1.20.0
gunicorn==20.1.0
idna==3.3
Pillow==9.0.1
psycopg2==2.9.3
psycopg2-binary==2.8.6
PyJWT==1.7.1
python-decouple==3.3
pytz==2021.3
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
whitenoise==6.0.0
